Turned test.js into folder, for clearer test separation
