+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "homepage": "https://github.com/finwo/supercop",
   "scripts": {
-    "test": "node test.js",
+    "test": "tape test/*.js",
     "prepublish": "npm test",
     "postpublish": "PACKAGE_VERSION=$(cat package.json | grep \\\"version\\\" | head -1 | awk -F: '{print $2}' | sed 's/[\",]//g' | tr -d '[[:space:]]') && npm deprecate \"supercop@<${PACKAGE_VERSION}\" \"Rolling release, please update to ${PACKAGE_VERSION}\"",
     "build": "make"
